requirements.txt
Django>=3.2.8,<3.3.0
djangorestframework>=3.12.4,<3.13.0
python-decouple==3.5
dj-database-url==0.5.0
celery>=5.1.2,<5.2.0
redis>=3.5.3,<3.6.0
django-filter==21.1
psycopg2-binary>=2.9.1,<2.10.0
drf-yasg>=1.20.0,<1.21.0